SF2: Handling ranges in preset zones better.
All preset generators are supposed to be "relative" modifications of the
instrument generators, but this makes no sense for ranges.  Previously, we
ignored ranges (key and velocity) in preset generators; now the instrument
ranges take precedence.  That may not be correct for the case where both the
preset and the instrument set a range.